requirements.txt
appdirs==1.4.3
arrow==0.15.4
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
bandit==1.5.1
bcrypt==3.1.7
beautifulsoup4==4.7.1
boto==2.49.0
boto3==1.9.121
botocore==1.12.121
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
configparser==3.7.4
cryptography==2.6.1
dj-database-url==0.4.2
Django==1.11.23
django-extensions==2.1.6
django-mathfilters==0.4.0
django-prettyjson==0.4.1
docutils==0.14
dropbox==9.3.0
enum34==1.1.6
futures==3.2.0
gitdb2==2.0.5
GitPython==2.1.11
gunicorn==19.9.0
idna==2.8
ipaddress==1.0.22
isort==4.3.16
jmespath==0.9.4
lazy-object-proxy==1.3.1
lockfile==0.12.2
lxml==4.3.4
mccabe==0.6.1
numpy==1.16.2
oauthlib==3.0.1
packaging==19.0
paramiko==2.6.0
pbr==5.1.3
psutil==5.6.1
psycopg2-binary==2.8.3
pushjack==1.6.0
pycparser==2.19
pyfcm==1.4.5
pylint==1.9.4
PyNaCl==1.3.0
pyparsing==2.3.1
python-dateutil==2.6.0
pytz==2018.9
PyYAML==5.1
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
s3transfer==0.2.0
singledispatch==3.4.0.3
six==1.12.0
smmap2==2.0.5
soupsieve==1.9.1
sphinx-me==0.3
standardjson==0.3.1
stevedore==1.30.1
typing==3.6.6
urllib3==1.24.3
whitenoise==4.1.2
wrapt==1.11.1
zipstream==1.1.4
